requirements.txt
asgiref==3.2.7
certifi==2020.4.5.2
cloudinary==1.21.1
dj-database-url==0.5.0
Django==3.0.7
django-cors-headers==3.3.0
gunicorn==20.0.4
imutils==0.5.3
numpy==1.18.5
opencv-contrib-python-headless==4.2.0.34
opencv-python-headless==4.2.0.34
Pillow==7.1.2
psycopg2-binary==2.8.5
pytesseract==0.3.4
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
